Move CRUD db operations to sep class
